+++ common/src/main/kotlin/com/project/common/data/responses/xp/XpTierResponse.kt
@@ -1,11 +1,13 @@
 package com.project.common.data.responses.xp
 
+import java.math.BigDecimal
+
 data class XpTierResponse(
     val id: Long,
     val minXp: Long,
     val maxXp: Long,
     val name: String,
-    val xpPerkMultiplier: Long,
+    val xpPerkMultiplier: BigDecimal,
     val xpPerNotification: Long,
     val xpPerPromotion: Long,
     val perkAmountPercentage: Long
